:sparkles: Integrate disk-prober with ecs-disk-manager
Now that the disk prober's functionality is almost complete, it is time
to integrate it into our ECS so that we can try it out in practice. As
soon as we've finished integrating basic disk probing with the ECS, we
should look into improving the support for LVM logical volumes and their
volume groups.

It seems to be integrating really well -- very little effort was
required to interface with the ECS. The ECS is able to provide
assocations in both ways for a device: associating parent devices with
child(ren) and children with their parent device(s).

Note that at this time the ECS is only available to gather information. I
haven't gotten to implementing the S in ECS yet. As soon as LVM and LUKS
device information gathering is complete, I can begin working on
integrating partition table management support.