:construction: Add lvmdbus1 & cryptsetup members
These crates do not exist, but they must if we want to avoid calling out
to the shell for every LVM and LUKS operation. **lvmdbus1** is the
official DBus API for managing LVM devices, and **cryptsetup** will be
interfacing with the **libcryptsetup.h** C API for managing LUKS devices.

Note that the LVM DBus API is the only stable API offered by the LVM
project at the moment. It would appear that they have deprecated the
**lvm2app.h** C API, leaving only the DBus API in its wake.