:sparkles: Complete cryptsetup device info methods
It is now possible to init a crypt device context to read information
from the crypt device's header, such as the name of the backing device
that is encrypted, the UUID of that device, the sector size, cipher
info, LUKS version type, and any offsets.

An example is included in the cryptsetup crate for dumping this
information.