:sparkles: LVM device info support
There is now enough information in the lvmdbus1 DBus bindings to
determine which PVs and LVs belong to which VGs, as well as the combined
size of a VG and the amount of free space is left.

In the future, I will add the DBus methods for managing these devices.