feat(lvm): Integrate LVM scanning into the ECS
LVM scanning via `lvmdbus1` is now fully feature-complete. It is now
possible to determine which devices are associated with what LVM volume
groups (VGs).

Devices are associated with VGs by either a physical volume (PV) or
logical volume (LV) association. A VG consists of one or more PVs, and
provides a virtual partitionable device for creating partitions on it as
LVs. Volume groups can therefore be treated similarly to a physical disk
with a partition table, although they do not exist as an actual device
on the system.

In addition to integrating LVM information into the ECS, I've refactored
the code to separate the scanning function to a system module. Some
additional convenience methods have also been added to iterate upon
specific types of devices, such as disks, partitions, LVs, and PVs.

There's also now a disk-ops sub-crate which will contain functions for
performing disk operations, and a few of these dummy functions are now
in the ECS disk manager to experiment with the API.