feat(manager): First Class Entities, Components, & Systems
BREAKING CHANGE: The disk manager structure now contains the following new fields:

- DiskEntities
- DiskComponents
- DiskSystems

The new `DiskSystems` structure holds systems which contain state.

All systems have been modified to no longer take the `DiskManager` as
a whole, but will take the `DiskEntities` and `DiskComponents` in
seperate mutable borrows, to increase the flexibility of borrowing
when a system in the world contains mutable state itself.

The `CreationSystem` is the first to be promoted to own state.